requirements.txt
Django==2.1.1
djangorestframework==3.8.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
selenium==3.14.1
urllib3==1.23
